+++ WorldCities/WorldCities/src/app/cities/cities.component.ts
@@ -21,7 +21,9 @@ export class CitiesComponent implements OnInit {
 
   defaultPageIndex: number = 0;
   defaultPageSize: number = 10;
-  public defaultSortColumn: string = "name";
+  defaultFilterColumn: string = "name";
+  filterQuery?: string;
+  public defaultSortColumn: string = "id";
   public defaultSortOrder: "asc" | "desc" = "asc";
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
@@ -33,10 +35,12 @@ export class CitiesComponent implements OnInit {
     this.loadData();
   }
 
-  loadData() {
+  loadData(query?: string) {
     var pageEvent = new PageEvent(); // generated a page event to call getData for the initial component load
     pageEvent.pageIndex = this.defaultPageIndex;
     pageEvent.pageSize = this.defaultPageSize;
+    this.filterQuery = query;
+    console.log(query);
     this.getData(pageEvent);
   }
 
@@ -48,6 +52,12 @@ export class CitiesComponent implements OnInit {
       .set("pageSize", event.pageSize.toString())
       .set("sortColumn", (this.sort) ? this.sort.active : this.defaultSortColumn)
       .set("sortOrder", (this.sort) ? this.sort.direction : this.defaultSortOrder);
+
+    if (this.filterQuery) {
+      params = params
+        .set("filterColumn", this.defaultFilterColumn)
+        .set("filterQuery", this.filterQuery)
+    }
 
     this.http.get<any>(url, { params })
       .subscribe(result => {